model-worker/requirements.txt
numpy==1.26.4
torch==2.1.0
torchvision==0.16.0
diffusers==0.24.0
transformers==4.38.2
huggingface_hub==0.20.3
accelerate==0.26.1
safetensors==0.4.2
Pillow==10.2.0
celery==5.3.6
redis==5.0.3
azure-storage-blob==12.19.1
python-dotenv==1.0.1
